refactor: event handler, service 2가지 클래스로만 구분하도록 수정
